+++ packages/vscode-extension/.eslintrc.js
@@ -5,7 +5,6 @@ module.exports = {
     },
     "extends": [
         "prettier",
-        "prettier/@typescript-eslint"
     ],
     "parser": "@typescript-eslint/parser",
     "parserOptions": {