+++ packages/vscode-extension/src/commands/UpdateObject.ts
@@ -11,6 +11,9 @@ import { ValidationResult } from '../types/ActionResult';
 import { FOLDERS } from '../ApplicationConstants';
 
 const COMMAND_NAME = 'updateobject';
+const CUSTOM_RECORD_PREFIX = 'customrecord';
+const INCLUDE_INSTANCES = 'includeinstances';
+
 const STATUS = {
 	SUCCESS: 'SUCCESS',
 	ERROR: 'ERROR',
@@ -31,19 +34,40 @@ export default class UpdateObject extends BaseAction {
 
 		const continueMessage = this.translationService.getMessage(ANSWERS.CONTINUE);
 		const cancelMessage = this.translationService.getMessage(ANSWERS.CANCEL);
-		const override = await window.showQuickPick([continueMessage, cancelMessage], {
-			placeHolder: this.translationService.getMessage(UPDATE_OBJECT.OVERRIDE, scriptId),
+		const yes = this.translationService.getMessage(ANSWERS.YES);
+		const no = this.translationService.getMessage(ANSWERS.NO);
+		const processCanceledMessage = this.translationService.getMessage(UPDATE_OBJECT.PROCESS_CANCELED);
+		const suiteCloudCommandOptions = { scriptid: [scriptId], includeinstances: '' };
+		let overwriteMessage = this.translationService.getMessage(UPDATE_OBJECT.OVERWRITE, scriptId);
+
+		if (this.isCustomRecord(scriptId)) {
+			const includeInstancesAnswer = await window.showQuickPick([yes, no], {
+				placeHolder: this.translationService.getMessage(UPDATE_OBJECT.QUESTIONS.INCLUDE_INSTANCES),
+				canPickMany: false,
+			});
+			if(!includeInstancesAnswer) {
+				this.messageService.showInformationMessage(processCanceledMessage);
+				return;
+			}
+			if (includeInstancesAnswer === yes) {
+				suiteCloudCommandOptions.includeinstances = INCLUDE_INSTANCES;
+				overwriteMessage = this.translationService.getMessage(UPDATE_OBJECT.OVERWRITE_INSTANCES, scriptId);
+			}
+		}
+
+		const overwriteObject = await window.showQuickPick([continueMessage, cancelMessage], {
+			placeHolder: overwriteMessage,
 			canPickMany: false,
 		});
 
-		if (!override || override === this.translationService.getMessage(ANSWERS.CANCEL)) {
-			this.messageService.showInformationMessage(this.translationService.getMessage(UPDATE_OBJECT.PROCESS_CANCELED));
+		if (!overwriteObject || overwriteObject === cancelMessage) {
+			this.messageService.showInformationMessage(processCanceledMessage);
 			return;
 		}
 
 		const commandMessage = this.translationService.getMessage(COMMAND.TRIGGERED, this.vscodeCommandName);
 		const statusBarMessage = this.translationService.getMessage(UPDATE_OBJECT.UPDATING);
-		const commandActionPromise = this.runSuiteCloudCommand({ scriptid: [scriptId] });
+		const commandActionPromise = this.runSuiteCloudCommand(suiteCloudCommandOptions);
 		this.messageService.showInformationMessage(commandMessage, statusBarMessage, commandActionPromise);
 
 		const actionResult = await commandActionPromise;
@@ -78,5 +102,9 @@ export default class UpdateObject extends BaseAction {
 		} catch (e: any) {
 			return this.unsuccessfulValidation(e.getErrorMessage());
 		}
+	}
+
+	private isCustomRecord(scriptid: string): boolean {
+		return scriptid.startsWith(CUSTOM_RECORD_PREFIX);
 	}
 }