+++ packages/vscode-extension/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 1.3.0
+- New **SuiteCloud: Validate Project** command
+- New **SuiteCloud: Manage Accounts** command
+- New **SuiteCloud: Compare Files** command
+- Include custom instances from the account for custom records in the SuiteCloud: Update Object command
+- Update to Oracle JDK version 17
+- Removal of outdated dependencies
+
 ## 1.2.0
 
 - Update SuiteCloud SDK dependency
@@ -23,7 +31,7 @@
 - Fix for browser-based auth for Fedora 34
 - Block extension commands while dependencies are not downloaded
 - Fix for feature dependencies incorrectly added in Manifest
-- Update Read Me documentation
+- Update the Read Me documentation
 - Minor bug fixes and improvements
 
 ## 1.0.2